DONE: - logger class added to UserService
TODO:
- AuthEntryPointJWT (exception handler) ?

+++ src/main/java/pl/joannaz/culturalcentrewieliszew/user/UserService.java
@@ -1,5 +1,6 @@
 package pl.joannaz.culturalcentrewieliszew.user;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -25,22 +26,36 @@ public class UserService implements UserDetailsService {
     private static final Logger logger = LoggerFactory.getLogger(UserService.class);
 
     public void addUser(User user) {
+        logger.info("Adding user {} {} to the database.", user.getFirstName(), user.getLastName());
         // set headshot
         user.setHeadshot(UserHelper.isFemale(user.getFirstName()) ? "assets/images/avatar4.svg" : "assets/images/avatar3.svg");
         this.userRepository.save(user);
     }
 
-    public Optional<User> findUserByUsername(String username) {return this.userRepository.findByUsername(username);}
-    public boolean existsByUsername(String username) {return this.userRepository.existsByUsername(username);}
+    public boolean existsByUsername(String username) {
+        logger.info("Checking if user with username: {} exists.", username);
+        return this.userRepository.existsByUsername(username);
+    }
+
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        logger.info("Fetching user with username: {}.", username);
         return this.userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found in the database."));
+                .orElseThrow(() -> {
+                    logger.error("User with {} not found.", username);
+                    return new UsernameNotFoundException(String.format("User with %s not found.", username));
+                });
     }
+
     public List<User> getChildren(UUID parentId) {
+        logger.info("Fetching children for user with id: {}.", parentId);
         return userRepository.findByParentId(parentId);
     }
+
     public User addChild(User parentUser, User childUser) {
+        logger.info("Adding {} {} as a child of user {} {}.", childUser.getFirstName(), childUser.getLastName(),
+                parentUser.getFirstName(), parentUser.getLastName());
+
         // Set the parent ID
         childUser.setParentId(parentUser.getId());
 
@@ -58,11 +73,19 @@ public class UserService implements UserDetailsService {
     }
 
     public User updateChild (User updatedChild) {
-        User originalChild = userRepository.findById(updatedChild.getId())
-                .orElseThrow(() -> new IllegalStateException(
-                        "Child with this id does not exist."
-                ));
+        logger.info("Updating child: {} {} of user with id {}.",
+                updatedChild.getFirstName(), updatedChild.getLastName(), updatedChild.getParentId());
 
+        logger.info("Fetching original user.");
+        User originalChild = userRepository.findById(updatedChild.getId())
+                .orElseThrow(() -> {
+                    logger.error("Error during fetching child: {} {}.", updatedChild.getFirstName(), updatedChild.getLastName());
+                    return new EntityNotFoundException(
+                            String.format("Child with id %s does not found.", updatedChild.getId())
+                    );
+                });
+
+        logger.info("Checking whether child's first name or last name has been changed.");
         if(!Objects.equals(originalChild.getFirstName(), updatedChild.getFirstName())
         || !Objects.equals(originalChild.getLastName(), updatedChild.getLastName())) {
             originalChild.setFirstName(updatedChild.getFirstName());
@@ -83,22 +106,34 @@ public class UserService implements UserDetailsService {
     }
 
     public User getUser (UUID userId) {
+        logger.info("Fetching user with id: {}.", userId);
         return userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalStateException(
-                        String.format("User with id %s does not exist.", userId)
-                ));
+                .orElseThrow(() -> {
+                    logger.error("Error during fetching user with id: {}.", userId);
+                    return new EntityNotFoundException(String.format("User with id %s not found.", userId));
+                });
     }
 
     public void deleteChild (UUID childId) { // corresponding rows in UserCourse are deleted automatically (removing child from lists of participants)
+        logger.info("Checking whether child with id: {} exists.", childId);
+        boolean childExists = userRepository.existsById(childId);
+        if (!childExists) {
+            logger.error("Child with id: {} does not exist.", childId);
+            throw new EntityNotFoundException(String.format("Child with id %s not found.", childId));
+        }
+
+        logger.info("Deleting child with id: {}", childId);
         userRepository.deleteById(childId);
     }
 
     public User getCurrentUser() {
+        logger.info("Fetching current user from Security Context Holder.");
         return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     }
 
     public List<CourseBasicInfo> getCoursesForUser(UUID userId) {
         User user = this.getUser(userId);
+        logger.info("Fetching courses for user with id: {}", userId);
         return user.getCourses().stream()
                 .map(UserCourse::getCourse)
                 .map(CourseBasicInfo::new)
@@ -106,42 +141,32 @@ public class UserService implements UserDetailsService {
     }
 
     @Transactional
-    public void addCourse(Long courseId) {
-            Course course = this.courseRepository.findById(courseId).orElseThrow(() ->
-                    new NoSuchElementException(String.format("Course with id %s not found.", courseId))
-            );
-//        CourseDetails courseDetails = courseDetailsRepository.findById(course.getId()).orElseThrow(() ->
-//                new NoSuchElementException(String.format("Course details with id %s not found.", courseId))
-//        );
-
-            if (course.getParticipants().size() < course.getMaxParticipantsNumber()) {
-                User currentUser = getCurrentUser();
-//            if (isAgeCorrect(currentUser.getDob(), courseDetails.getMinAge(), courseDetails.getMaxAge()))
-                currentUser.addCourse(course);
-//            else throw new RuntimeException("Sorry, your age does not fit within the age range of this course.");
-            } else {
-                throw new RuntimeException("Sorry, there is no more available slots for this course.");
-            }
-    }
-
-    @Transactional
     public void joinCourse(Long courseId, UUID userId) {
         // get course
-        Course course = this.courseRepository.findById(courseId).orElseThrow(() ->
-                new NoSuchElementException(String.format("Course with id %s not found.", courseId))
-        );
+        logger.info("Fetching course with id: {}", courseId);
+        Course course = this.courseRepository.findById(courseId).orElseThrow(() -> {
+            logger.error("Error during fetching course with id: {}", courseId);
+            return new EntityNotFoundException(String.format("Course with id %s not found.", courseId));
+        });
+
         // get course details to check if user age match minAge and maxAge values
-        CourseDetails courseDetails = courseDetailsRepository.findById(courseId).orElseThrow(() ->
-                new NoSuchElementException(
-                        String.format("Course Details with id %s not found.", courseId)
-                ));
+        logger.info("Fetching course details with id: {} for {} course.", courseId, course.getName());
+        CourseDetails courseDetails = courseDetailsRepository.findById(courseId).orElseThrow(() -> {
+            logger.error("Error during fetching course details with id: {} for {} course.", courseId, course.getName());
+            return new EntityNotFoundException(String.format("Course details with id %s not found.", courseId));
+        });
+
         // get user
-        User user = this.userRepository.findById(userId).orElseThrow(() ->
-                new NoSuchElementException(String.format("User with id %s not found.", userId))
-        );
+        logger.info("Fetching user with id: {}.", userId);
+        User user = this.userRepository.findById(userId).orElseThrow(() -> {
+            logger.error("Error during fetching user with id: {}", userId);
+            return new EntityNotFoundException(String.format("User with id %s not found.", userId));
+        });
 
         // check if there is free slot for new participant
+        logger.info("Checking whether there are free slots for {} course.", course.getName());
         if (course.getParticipants().size() < course.getMaxParticipantsNumber()) {
+            logger.info("Validating age");
             // check if user age match minAge and maxAge values
             if (isAgeCorrect(user.getDob(), courseDetails.getMinAge(), courseDetails.getMaxAge())) {
                 // join course
@@ -160,10 +185,10 @@ public class UserService implements UserDetailsService {
     @Transactional
     public void removeCourse(Long courseId, UUID userId) {
         Course course = this.courseRepository.findById(courseId).orElseThrow(() ->
-                new NoSuchElementException(String.format("Course with id %s not found.", courseId))
+                new EntityNotFoundException(String.format("Course with id %s not found.", courseId))
         );
         User user = this.userRepository.findById(userId).orElseThrow(() ->
-                new NoSuchElementException(String.format("User with id %s not found.", userId))
+                new EntityNotFoundException(String.format("User with id %s not found.", userId))
         );
         user.removeCourse(course);
     }
@@ -182,7 +207,7 @@ public class UserService implements UserDetailsService {
 
     public EmployeeProfile getEmployeeById(UUID id) {
         User employee = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalStateException(
+                .orElseThrow(() -> new EntityNotFoundException(
                         String.format("Employee with id %s does not exist.", id)
                 ));
         return new EmployeeProfile(id, employee.getFirstName(), employee.getLastName(),
@@ -192,7 +217,7 @@ public class UserService implements UserDetailsService {
     @Transactional
     public void deleteEmployee(UUID employeeId) {
         User employee = userRepository.findById(employeeId)
-                .orElseThrow(() -> new IllegalStateException(
+                .orElseThrow(() -> new EntityNotFoundException(
                         String.format("Employee with id %s does not exist.", employeeId)
                 ));
         if(employee.getPosition().equals("Teacher")) { // get teacher courses and set teacher value to null
@@ -210,7 +235,7 @@ public class UserService implements UserDetailsService {
     @Transactional
     public User updateEmployee(UserDTO updatedEmployee) {
         User originalEmployee = userRepository.findById(updatedEmployee.getId())
-                .orElseThrow(() -> new IllegalStateException(
+                .orElseThrow(() -> new EntityNotFoundException(
                         "Employee with this id does not exist."
                 ));
         originalEmployee.setFirstName(updatedEmployee.getFirstName());
